requirements.txt
fastapi==0.88.0
pymongo==4.3.3
uvicorn==0.20.0
python-decouple==3.6
